requirements.txt
beautifulsoup4==4.11.2
discord.py==2.1.0
fuzzywuzzy==0.18.0
pykakasi==2.2.1
pytesseract==0.3.10
requests==2.22.0
